move layout to models
